metrics-writer/requirements.txt
SQLAlchemy==1.4.41
google-cloud-monitoring==2.11.2
PyMySQL==1.0.2